requirements.txt
anaconda==0.0.1.1
appnope==0.1.0
asgiref==3.2.3
astroid==2.2.5
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
bleach==3.1.4
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.5
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-oauth-toolkit==1.3.2
django-photologue==3.9
django-sortedm2m==1.5.0
djangorestframework==3.11.0
entrypoints==0.3
ExifRead==2.1.2
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.17
jedi==0.16.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.3.1
libsass==0.19.4
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.3
oauthlib==3.1.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.3.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
requests==2.23.0
Send2Trash==1.5.0
six==1.14.0
SQLAlchemy==1.3.16
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.3.4
urllib3==1.25.9
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.1
zipp==3.1.0
